stormcloud/plugins/hms_grid/requirements.txt
affine==2.4.0
attrs==23.1.0
boto3==1.33.8
botocore==1.33.8
certifi==2023.11.17
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
GDAL==3.5.1
gitdb==4.0.11
GitPython==3.1.37
jmespath==1.0.1
numpy==1.24.3
pandas==2.0.3
papipyplug==2023.10.3
# pydsstools==2.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rasterio==1.3.9
s3transfer==0.8.2
shapely==2.0.2
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
tzdata==2023.3
urllib3==1.26.18